design: 승인 상세를 페이지에서 모달로 바꾼다 #239
